Created new class and added createUsers in to new test class
